move reduceQueryPropsDecorator to another file
